dock button in header, silence + connection buttons in menu
